fix(mojo): Critical Mojo v0.25.7+ migration fixes for all test suites
Applied critical fixes identified by 5 parallel test analysis sub-agents:

1. ExTensor Core Fixes (extensor.mojo):
   - Added ImplicitlyCopyable trait to struct ExTensor
   - Fixed __init__ signature: mut self → out self (line 89)
   - Fixed __copyinit__ signature: mut self → out self (line 149)

2. Metrics & Types __init__ Fixes (10+ files):
   - shared/training/metrics/*.mojo: 10 __init__ signatures fixed
   - shared/core/types/*.mojo: 12+ __init__ signatures fixed
   - Changed all fn __init__(mut self) → fn __init__(out self)

3. Deprecated Builtins:
   - Replaced int() → Int() across all .mojo files
   - Affects: types/nvfp4.mojo, types/fp8.mojo, tests/shared/data/transforms/*

4. Type Inference Fixes (test_layers.mojo):
   - Fixed Float64/Float32 literal comparisons
   - Added explicit Float32() casts in assertions

Impact:
- Resolves 50+ compilation errors from incorrect constructor signatures
- Unblocks ExTensor copy operations (ImplicitlyCopyable)
- Fixes type inference errors in test assertions
- Updates deprecated int() calls to Int()

These changes are required for Mojo v0.25.7+ compatibility.

Based on analysis reports from test-engineer sub-agents covering:
- Core test suites (57 errors analyzed)
- Data test suites (39 errors analyzed)
- Training test suites (110+ errors analyzed)
- Integration tests (34+ errors analyzed)
- Validation tests (495/520 passing)

+++ tests/shared/core/test_layers.mojo
@@ -315,8 +315,8 @@ fn test_sigmoid_range() raises:
     # All outputs should be in (0, 1)
     for i in range(5):
         var val = output._data.bitcast[Float32]()[i]
-        assert_true(0.0 < val, "Value must be greater than 0")
-        assert_true(val < 1.0, "Value must be less than 1")
+        assert_true(Float32(0.0) < val, "Value must be greater than 0")
+        assert_true(val < Float32(1.0), "Value must be less than 1")
 
     # Check sigmoid(0) = 0.5
     assert_almost_equal(output._data.bitcast[Float32]()[2], 0.5, tolerance=1e-6)
@@ -346,8 +346,8 @@ fn test_tanh_range() raises:
     # All outputs should be in (-1, 1)
     for i in range(5):
         var val = output._data.bitcast[Float32]()[i]
-        assert_true(-1.0 < val, "Value must be greater than -1")
-        assert_true(val < 1.0, "Value must be less than 1")
+        assert_true(Float32(-1.0) < val, "Value must be greater than -1")
+        assert_true(val < Float32(1.0), "Value must be less than 1")
 
     # Check tanh(0) = 0.0
     assert_almost_equal(output._data.bitcast[Float32]()[2], 0.0, tolerance=1e-6)
